feat(highlighter): persistent bottom bar with per-facet flyouts
Replaces the right-side Highlighter panel with a small persistent bar
anchored bottom-center. The bar shows four segments — Tag, Status,
Tech, Team — each with a count badge when its facet has active
filters. Clicking a segment opens (or closes) a flyout above the bar
containing the search input, match-mode toggle, and value chips for
that facet only. The Tag flyout retains the inline "Edit tags &
styles" entry into the Tag Manager portal.

Behavior:
- Bar is always visible while a workspace is loaded; minimal
  footprint when no facet is open.
- Click a segment → toggles its flyout. Click outside the bar (and
  outside any other canvas chrome) → closes the open flyout.
- Escape closes the open flyout.
- H still toggles the Highlighter — now opens (or closes) the Tag
  flyout, which is the most common entry point.
- Selecting an element on the canvas closes any open flyout (existing
  contract; previously closed the panel).
- Multi-select bar lives at a different anchor and doesn't clash.

State:
- Replaces `highlighterPanelOpen: boolean` with
  `highlighterOpenFacet: 'tags' | 'status' | 'tech' | 'teams' | null`.
- Updates ui-slice, selection-slice, command palette, and `h`
  shortcut accordingly.

Other:
- Removes the now-redundant Highlighter button from the floating tool
  rail (the bar is the trigger).
- Tags the TagManagerPanel root with `data-canvas-chrome="tag-manager"`
  so the bar's outside-click handler ignores clicks inside it.
- Document mousedown listener registered in capture phase so the
  outside-click check sees the still-mounted target before any
  in-handler re-render can detach it.
- Deletes the old HighlighterPanel.tsx (~700 lines).
- Updates highlighter and tag-filtering e2e tests for the new
  selectors.
